CARBON-15564: Adding org.wso2.carbon.kernel.utils.UtilsTest
